merge: 오늘 머지한 로그인 부분 재결합
프론트 작업과 결합시켰다. 아직 DB작업 필요

Related to: #42 #2
