+++ client/src/components/LoginPage/LoginBox.tsx
@@ -1,16 +1,22 @@
 import React from 'react';
 import styled from 'styled-components';
 import mainLogo from '../../images/main-logo.png';
-import githubLogo from '../../images/githubLogo.png';
 import { GithubLoginButton } from 'react-social-login-buttons';
 
 const LoginBox = () => {
+  const loginGithub = () => {
+    fetch('/oauth/login')
+      .then((res) => res.json())
+      .then((loginLink) => {
+        window.location.href = loginLink;
+      });
+  };
   return (
     <Box>
       <Wrapper>
         <MainLogoElement />
         <Text>개발자라면 Github 아이디는 가지고 계시죠?</Text>
-        <GithubLoginButton onClick={() => alert('Hello')} />
+        <GithubLoginButton onClick={loginGithub} />
       </Wrapper>
     </Box>
   );
